transport buttons functional
The functionality of the transport control model buttons is implemented and operating correctly for continuous and discrete sweeps on the pan axis.  Control of the positioner, specifically control of packet transmission, decoupled from measurment control and currently implemented inside of the MyMainWindow object.  This is a slight problem in terms of performance of the window if moving the window around, but its not bad - just kinda choppy.  This is a temporary solution for the alpha demo, mainly to facilitate correct MeasurementCtrl behavior when combined with transport control buttons; intent is to work on moving the positioner into a worker object to be controlled by QThreadPool - just not entirely clear on specifics of how to do this yet, will work on figuring it out over the next couple weeks. Also, there is some temporary stuff added to the UI, need to get with Austin to discuss what needs to stay so he can come up with permenant solution.